+++ src/main/java/com/project/jpa/repository/BookRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.project.jpa.model.Book;
 
-public interface BookRepository extends JpaRepository<Book,Integer> {
+public interface BookRepository extends JpaRepository<Book,Long> {
 	
 	List<Book> findByTitle(String name);
 	Book findOneByTitle(String name);
@@ -56,89 +56,89 @@ public interface BookRepository extends JpaRepository<Book,Integer> {
 	// Find within a room ("...WHERE r.id = :id" would also work, but I did it the other way just to prove to myself I could)
 	
 	@Query("select b from Room r JOIN r.books b WHERE b.room.id = :id AND lower(b.title) LIKE lower(concat('%', :t,'%'))")
-	List<Book> findTitleInRoom(@Param("t") String title, @Param("id") int id);
+	List<Book> findTitleInRoom(@Param("t") String title, @Param("id") long id);
 	
 	@Query("select b from Room r JOIN r.books b WHERE b.room.id = :id AND lower(b.genre) LIKE lower(concat('%', :g,'%'))")
-	List<Book> findGenreInRoom(@Param("g") String genre, @Param("id") int id);
+	List<Book> findGenreInRoom(@Param("g") String genre, @Param("id") long id);
 	
 	@Query("select b from Room r JOIN r.books b WHERE b.room.id = :id AND lower(b.title) LIKE lower(concat('%', :gt,'%')) OR lower(b.genre) LIKE lower(concat('%', :gt,'%'))")
-	List<Book> findGenreAndTitleInRoom(@Param("gt") String genretitle, @Param("id") int id);
+	List<Book> findGenreAndTitleInRoom(@Param("gt") String genretitle, @Param("id") long id);
 	
 	@Query("select b from Room r JOIN r.books b WHERE b.room.id = :id AND lower(b.more) LIKE lower(concat('%', :a,'%')) OR lower(b.title) LIKE lower(concat('%', :a,'%')) OR lower(b.genre) LIKE lower(concat('%', :a,'%')) OR lower(b.author) LIKE lower(concat('%', :a,'%'))")
-	List<Book> findAllInRoom(@Param("a") String all, @Param("id") int id);
+	List<Book> findAllInRoom(@Param("a") String all, @Param("id") long id);
 	
 	@Query("select b from Room r JOIN r.books b WHERE b.room.id = :id AND lower(b.color) LIKE lower(concat('%', :c,'%'))")
-	List<Book> findColorInRoom(@Param("c") String color, @Param("id") int id);
+	List<Book> findColorInRoom(@Param("c") String color, @Param("id") long id);
 	
 	@Query("select b from Room r JOIN r.books b WHERE b.room.id = :id AND lower(b.author) LIKE lower(concat('%', :a,'%'))")
-	List<Book> findAuthorInRoom(@Param("a") String author, @Param("id") int id);
+	List<Book> findAuthorInRoom(@Param("a") String author, @Param("id") long id);
 	
 	@Query("select b from Room r JOIN r.books b WHERE b.room.id = :id AND lower(b.more) LIKE lower(concat('%', :m,'%'))")
-	List<Book> findMoreInRoom(@Param("m") String more, @Param("id") int id);
+	List<Book> findMoreInRoom(@Param("m") String more, @Param("id") long id);
 	
 	@Query("select b from Room r JOIN r.books b WHERE b.room.id = :id AND b.pdate < :low AND b.pdate > :high")
-	List<Book> findPublishDateInRoom(@Param("low") int low, @Param("high") int high, @Param("id") int id);
+	List<Book> findPublishDateInRoom(@Param("low") int low, @Param("high") int high, @Param("id") long id);
 	
 	@Query("select b from Room r JOIN r.books b WHERE b.room.id = :id AND b.pages < :low AND b.pages > :high")
-	List<Book> findPagesInRoom(@Param("low") int low, @Param("high") int high, @Param("id") int id);
+	List<Book> findPagesInRoom(@Param("low") int low, @Param("high") int high, @Param("id") long id);
 	
 	// Find within a bookcase
 	
 	@Query("select b from Bookcase bk JOIN bk.books b WHERE b.bookcase.id = :id AND lower(b.title) LIKE lower(concat('%', :t,'%'))")
-	List<Book> findTitleInBookcase(@Param("t") String title, @Param("id") int id);
+	List<Book> findTitleInBookcase(@Param("t") String title, @Param("id") long id);
 	
 	@Query("select b from Bookcase bk JOIN bk.books b WHERE b.bookcase.id = :id AND lower(b.title) LIKE lower(concat('%', :g,'%'))")
-	List<Book> findGenreInBookcase(@Param("g") String genre, @Param("id") int id);
+	List<Book> findGenreInBookcase(@Param("g") String genre, @Param("id") long id);
 	
 	@Query("select b from Bookcase bk JOIN bk.books b WHERE b.bookcase.id = :id AND lower(b.title) LIKE lower(concat('%', :gt,'%')) OR lower(b.genre) LIKE lower(concat('%', :gt,'%'))")
-	List<Book> findGenreAndTitleInBookcase(@Param("gt") String genretitle, @Param("id") int id);
+	List<Book> findGenreAndTitleInBookcase(@Param("gt") String genretitle, @Param("id") long id);
 	
 	@Query("select b from Bookcase bk JOIN bk.books b WHERE b.bookcase.id = :id AND lower(b.more) LIKE lower(concat('%', :a,'%')) OR lower(b.title) LIKE lower(concat('%', :a,'%')) OR lower(b.genre) LIKE lower(concat('%', :a,'%')) OR lower(b.author) LIKE lower(concat('%', :a,'%'))")
-	List<Book> findAllInBookcase(@Param("a") String all, @Param("id") int id);
+	List<Book> findAllInBookcase(@Param("a") String all, @Param("id") long id);
 	
 	@Query("select b from Bookcase bk JOIN bk.books b WHERE b.bookcase.id = :id AND lower(b.color) LIKE lower(concat('%', :c,'%'))")
-	List<Book> findColorInBookcase(@Param("c") String color, @Param("id") int id);
+	List<Book> findColorInBookcase(@Param("c") String color, @Param("id") long id);
 	
 	@Query("select b from Bookcase bk JOIN bk.books b WHERE b.bookcase.id = :id AND lower(b.author) LIKE lower(concat('%', :a,'%'))")
-	List<Book> findAuthorInBookcase(@Param("a") String author, @Param("id") int id);
+	List<Book> findAuthorInBookcase(@Param("a") String author, @Param("id") long id);
 	
 	@Query("select b from Bookcase bk JOIN bk.books b WHERE b.bookcase.id = :id AND lower(b.more) LIKE lower(concat('%', :m,'%'))")
-	List<Book> findMoreInBookcase(@Param("m") String more, @Param("id") int id);
+	List<Book> findMoreInBookcase(@Param("m") String more, @Param("id") long id);
 	
 	@Query("select b from Bookcase bk JOIN bk.books b WHERE b.bookcase.id = :id AND b.pdate < :low AND b.pdate > :high")
-	List<Book> findPublishDateInBookcase(@Param("low") int low, @Param("high") int high, @Param("id") int id);
+	List<Book> findPublishDateInBookcase(@Param("low") int low, @Param("high") int high, @Param("id") long id);
 	
 	@Query("select b from Bookcase bk JOIN bk.books b WHERE b.bookcase.id = :id AND b.pages < :low AND b.pages > :high")
-	List<Book> findPagesInBookcase(@Param("low") int low, @Param("high") int high, @Param("id") int id);
+	List<Book> findPagesInBookcase(@Param("low") int low, @Param("high") int high, @Param("id") long id);
 	
 	// Find within a shelf
 	
 	@Query("select b from Shelf sh JOIN sh.books b WHERE b.shelf.id = :id AND lower(b.title) LIKE lower(concat('%', :t,'%'))")
-	List<Book> findTitleInShelf(@Param("t") String title, @Param("id") int id);
+	List<Book> findTitleInShelf(@Param("t") String title, @Param("id") long id);
 	
 	@Query("select b from Shelf sh JOIN sh.books b WHERE b.shelf.id = :id AND lower(b.title) LIKE lower(concat('%', :g,'%'))")
-	List<Book> findGenreInShelf(@Param("g") String genre, @Param("id") int id);
+	List<Book> findGenreInShelf(@Param("g") String genre, @Param("id") long id);
 	
 	@Query("select b from Shelf sh JOIN sh.books b WHERE b.shelf.id = :id AND lower(b.title) LIKE lower(concat('%', :gt,'%')) OR lower(b.genre) LIKE lower(concat('%', :gt,'%'))")
-	List<Book> findGenreAndTitleInShelf(@Param("gt") String genretitle, @Param("id") int id);
+	List<Book> findGenreAndTitleInShelf(@Param("gt") String genretitle, @Param("id") long id);
 	
 	@Query("select b from Shelf sh JOIN sh.books b WHERE b.shelf.id = :id AND lower(b.more) LIKE lower(concat('%', :a,'%')) OR lower(b.title) LIKE lower(concat('%', :a,'%')) OR lower(b.genre) LIKE lower(concat('%', :a,'%')) OR lower(b.author) LIKE lower(concat('%', :a,'%'))")
-	List<Book> findAllInShelf(@Param("a") String all, @Param("id") int id);
+	List<Book> findAllInShelf(@Param("a") String all, @Param("id") long id);
 	
 	@Query("select b from Shelf sh JOIN sh.books b WHERE b.shelf.id = :id AND lower(b.color) LIKE lower(concat('%', :c,'%'))")
-	List<Book> findColorInShelf(@Param("c") String color, @Param("id") int id);
+	List<Book> findColorInShelf(@Param("c") String color, @Param("id") long id);
 	
 	@Query("select b from Shelf sh JOIN sh.books b WHERE b.shelf.id = :id AND lower(b.author) LIKE lower(concat('%', :a,'%'))")
-	List<Book> findAuthorInShelf(@Param("a") String author, @Param("id") int id);
+	List<Book> findAuthorInShelf(@Param("a") String author, @Param("id") long id);
 	
 	@Query("select b from Shelf sh JOIN sh.books b WHERE b.shelf.id = :id AND lower(b.more) LIKE lower(concat('%', :m,'%'))")
-	List<Book> findMoreInShelf(@Param("m") String more, @Param("id") int id);
+	List<Book> findMoreInShelf(@Param("m") String more, @Param("id") long id);
 	
 	@Query("select b from Shelf sh JOIN sh.books b WHERE b.shelf.id = :id AND b.pdate < :low AND b.pdate > :high")
-	List<Book> findPublishDateInShelf(@Param("low") int low, @Param("high") int high, @Param("id") int id);
+	List<Book> findPublishDateInShelf(@Param("low") int low, @Param("high") int high, @Param("id") long id);
 	
 	@Query("select b from Shelf sh JOIN sh.books b WHERE b.shelf.id = :id AND b.pages < :low AND b.pages > :high")
-	List<Book> findPagesInShelf(@Param("low") int low, @Param("high") int high, @Param("id") int id);
+	List<Book> findPagesInShelf(@Param("low") int low, @Param("high") int high, @Param("id") long id);
 	
 }
